ecal/tests/integration_tests/requirements.txt
# Robot Framework Core
robotframework

# Docker Support
docker

# Robot Framework Libraries
robotframework-pythonlibcore